+++ types/node_windows_x64/aa.d.ts
@@ -1,1 +1,3 @@
 export declare function aa(asmString: string, rcx?: number): number;
+
+export declare function asmBytes(asmString: string): ArrayBuffer;